+++ src/containers/App/theme/index.js
@@ -1,0 +1,2 @@
+export { default as GlobalStyle } from './GlobalStyle';
+export { default } from './ThemeProvider';
