--- src/containers/App/theme/primary.js
@@ -1,18 +1,0 @@
-const theme = {
-  color: {
-    white: '#fefefe',
-    silver: '#ccc',
-    gray: '#888',
-    mineShaft: '#333',
-    mariner: '#2472be',
-    pictonBlue: '#23a5f0',
-  },
-
-  fonts: {
-    default: "'PT Sans', 'Helvetica Neue', Helvetica, Arial, sans-serif",
-    serif:
-      "'PT Serif', Palatino, 'Palatino Linotype', 'Palatino LT STD', 'Book Antiqua', Georgia, serif",
-  },
-};
-
-export default theme;
